+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.3.0] - 2023-05-12
+
+- Allow config via command line args
+- Allow combining config file with command line args
+
 ## [0.2.2] - 2025-04-20
 
 - Make files and folders optional
@@ -15,6 +20,7 @@
 
 - Initial release
 
+[0.3.0]: https://github.com/wenxuanjun/oib/releases/tag/v0.3.0
 [0.2.2]: https://github.com/wenxuanjun/oib/releases/tag/v0.2.2
 [0.2.1]: https://github.com/wenxuanjun/oib/releases/tag/v0.2.1
 [0.2.0]: https://github.com/wenxuanjun/oib/releases/tag/v0.2.0